issue #626 Rename the file according to its contents.
